+++ DriverDTS/src/main/java/connections/IDriverQuery.java
@@ -1,6 +1,6 @@
 package connections;
 
-import aidas.usersystem.dto.UserDto;
+import aidas.userservice.dto.UserDto;
 
 /**
  * Created by geh on 22-4-14.
@@ -9,7 +9,7 @@ public interface IDriverQuery
 {
     UserDto authenticate(String userId, String password);
 
-    Integer getLaneCount();
+    Long getLaneCount();
 
     Long getEdgeCount();
 }
